+++ src/onegov.feriennet/onegov/feriennet/assets/js/print.js
@@ -1,7 +1,13 @@
-$('.print[data-print-selector]').click(function() {
+$('.print[data-print-selector]').click(function(e) {
     var el = $(this);
 
-    $(el.data('print-selector')).printThis({
-        'pageTitle': el.data('print-title') || ''
-    });
+    if ($(el).data('is-print-hooked') !== true) {
+        $(el.data('print-selector')).printThis({
+            'pageTitle': el.data('print-title') || '',
+            'loadCSS': el.data('print-css') || ''
+        });
+        $(el).data('is-print-hooked', true);
+    }
+
+    e.preventDefault();
 });